refactor: split web/server.go into 6 handler files + infrastructure
